CRUD de productos finalizado

+++ app/Http/Controllers/WebController/ProductController.php
@@ -3,16 +3,19 @@
 namespace App\Http\Controllers\WebController;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Product\ProductRequest;
+use App\Http\Requests\Product\ProductUpdateRequest;
 use App\Models\Product;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 use Illuminate\Routing\Redirector;
+use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
 {
+
     public function __construct()
     {
         $this->middleware('auth');
@@ -29,10 +32,17 @@ class ProductController extends Controller
         return view('components.product.create');
     }
 
-    public function store(Request $request): Redirector|Application|RedirectResponse
+    public function store(ProductRequest $request): Redirector|Application|RedirectResponse
     {
-        $product = new Product($request->all());
+        $data = request()->except('_token');
+
+        if ($request->hasFile('image')) {
+            $data['image'] = $request->file('image')->store('uploads', 'public');
+        }
+
+        $product = new Product($data);
         $product->save();
+
         return redirect('product/management')->with('message', 'Producto agregado correctamente');
     }
 
@@ -41,16 +51,27 @@ class ProductController extends Controller
         return view('components.product.edit', compact('product'));
     }
 
-    public function update(Product $product, Request $request): Redirector|Application|RedirectResponse
+    public function update(Product $product, ProductUpdateRequest $request): Redirector|Application|RedirectResponse
     {
-        $product->update($request->all());
+        $data = request()->except(['_token', '_method']);
+
+        if ($request->hasFile('image')) {
+            Storage::delete('public/' . $product->getAttribute('image'));
+            $data['image'] = $request->file('image')->store('uploads', 'public');
+        }
+
+        $product->update($data);
         $product->save();
+
         return redirect('product/management')->with('message', 'Producto actualizado correctamente');
     }
 
     public function destroy(Product $product): Redirector|Application|RedirectResponse
     {
-        $product->delete();
+        if (Storage::delete('public/' . $product->getAttribute('image'))) {
+            $product->delete();
+        }
+
         return redirect('product/management')->with('message', 'Producto eliminado correctamente');
     }
